photoric/requirements.txt
python3.6
flask
flask_authorize
flask_login
flask_session
flask_sqlalchemy
flask_wtf
datetime
dotenv
IPTCinfo3
werkzeug.security
email-validator
pillow